+++ components/src/docs/HowTo.story.vue
@@ -1,0 +1,35 @@
+<template>
+  <Story title="Components" docs-only icon="la:book-open" group="docs" />
+</template>
+
+<docs lang="md">
+
+![](/components.svg)
+
+# Gun-Vue is a library of reusable components for Gun built with Vue.
+
+It enables eager web developers to build p2p apps for their communities all over the world, free and open source.
+
+## How to install?
+
+Gun-Vue is distributed as a regular NPM package and may be used right from your Vue templates. 
+
+1. You may start by [creating a new Vue project](https://vuejs.org/guide/quick-start.html).
+
+```shell
+pnpm init vue@latest
+cd <your-project-name>
+pnpm install
+pnpm run dev
+```
+2. Add a new Single File Component (SFC), or add the code to an existing Route/Component
+
+```js
+
+
+```
+
+</docs>
+
+<style scoped>
+</style>