+++ components/src/index.js
@@ -2,19 +2,18 @@ import "virtual:windi.css";
 import "./styles/index.css";
 import "./styles/transitions.css";
 
-export * from './account/index.js'
-export * from './chat/index.js'
-export * from './dict/index.js'
-export * from './form/index.js'
-export * from './gift/index.js'
-export * from './post/index.js'
-export * from './qr/index.js'
-export * from './room/index.js'
-export * from './space/index.js'
-export * from './ui/index.js'
-export * from './user/index.js'
-export * from './util/index.js'
+import * as components from './all.js'
 
+export const GunVuePlugin = {
+  install(Vue) {
+    for (const prop in components) {
+      Vue.component(prop, components[prop])
+    }
+  }
+}
+
+export * from './all.js'
 export * from "#composables";
 
 export { default as FloatingVue } from 'floating-vue'
+