+++ composables/jsdoc2md/docs.sh
@@ -1,1 +1,1 @@
-jsdoc2md --partial sig-name.hbs --helper helpers.js --template docs.hbs --files ../src/**/*.js > ../../components/src/composables/Composables.story.vue
+jsdoc2md --partial sig-name.hbs --helper helpers.js --template docs.hbs --files ../src/**/*.js > ../../components/src/docs/Composables.story.vue