+++ composables/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gun-vue/composables
 
+## 0.10.1
+
+### Patch Changes
+
+- Vue plugin components export
+
 ## 0.10.0
 
 ### Minor Changes